# api/openapi.yaml
openapi: 3.1.0
info:
  title: Chat Server API
  version: 1.0.0
  description: |
    API contract for the Go/PostgreSQL chat server.
    This contract is expanded alongside the implementation; keep generated
    code synchronized. Web client requirements: docs/WEB_CLIENT.md.
    Decisions referenced here: docs/DECISIONS.md (ADR-009 receipts, ADR-010
    contacts, ADR-011 avatars, ADR-013 preferences, ADR-014 CLI accounts).
servers:
  - url: /api/v1
...
paths:
...
  /users/me:
    get:
      tags: [users]
      security: [{ bearerAuth: [] }]
      operationId: getMe
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      tags: [users]
      security: [{ bearerAuth: [] }]
      operationId: updateMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
  /users/me/preferences:
    get:
      tags: [users]
      security: [{ bearerAuth: [] }]
      operationId: getPreferences
      responses:
        "200":
          description: Current preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
    patch:
      tags: [users]
      security: [{ bearerAuth: [] }]
      operationId: updatePreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferencesRequest"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
  /users/me/notifications:
    get:
      tags: [users]
      security: [{ bearerAuth: [] }]
      operationId: listNotificationModes
      description: >
        Caller's explicit notification modes. rows with room_id "" are the
        default; others are room overrides (docs/REQUIREMENTS.md).
      responses:
        "200":
          description: Notification modes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationMode"
    post:
      tags: [users]
      security: [{ bearerAuth: [] }]
      operationId: setNotificationMode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetNotificationModeRequest"
      responses:
        "204":
          description: Mode stored
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /rooms/{roomId}/archived:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    put:
      tags: [rooms]
      security: [{ bearerAuth: [] }]
      operationId: setMemberArchive
      description: >
        Per-member conversation archive (web client's archived view,
        docs/CLIENT_WEB.md). Distinct from the admin room-level archive
        (`DELETE /rooms/{roomId}`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberArchiveRequest"
      responses:
        "204":
          description: Updated
...
  /messages/{messageId}/reactions/{emoji}:
    parameters:
      - $ref: "#/components/parameters/MessageId"
      - in: path
        name: emoji
        required: true
        schema:
          type: string
    delete:
      tags: [messages]
      security: [{ bearerAuth: [] }]
      operationId: unreact
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RoomId:
      in: path
      name: roomId
      required: true
      schema:
        type: string
    MessageId:
      in: path
      name: messageId
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable message.
...
    User:
      type: object
      required: [id, username, display_name, role, avatar_url]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum: [user, moderator, admin]
        email:
          type: string
          nullable: true
          description: Present only on self payloads (GET/PATCH /users/me, login).
        timezone:
          type: string
          nullable: true
          description: IANA tz name for client-side rendering; UTC RFC 3339 on the wire.
        avatar_url:
          type: string
          nullable: true
          description: >
            Resolved avatar URL (presigned attachment URL, or server-side
            Gravatar fallback, or null for client-side monogram) — ADR-011.
        last_seen_at:
          type: string
          format: date-time
          nullable: true
          description: >
            Null unless the target user's lastSeen preference allows, ADR-013.
    UpdateUserRequest:
      type: object
      properties:
        display_name:
          type: string
        email:
          type: string
          nullable: true
        avatar_attachment_id:
          type: string
          nullable: true
          description: Set from an attachment created via the upload flow (ADR-011).
    Preferences:
      type: object
      required: [last_seen_visible, read_receipts, allow_group_invites, allow_private_messages]
      properties:
        last_seen_visible:
          type: boolean
        read_receipts:
          type: boolean
        allow_group_invites:
          type: boolean
        allow_private_messages:
          type: boolean
    UpdatePreferencesRequest:
      description: Partial update; omitted fields keep current values.
      type: object
      properties:
        last_seen_visible:
          type: boolean
        read_receipts:
          type: boolean
        allow_group_invites:
          type: boolean
        allow_private_messages:
          type: boolean
      minProperties: 1
...
    NotificationMode:
      type: object
      required: [room_id, mode]
      properties:
        room_id:
          type: string
          description: "Room id for overrides; \"\" for the user default row."
        mode:
          type: string
          enum: [all, mentions, directs, never]
...
    SetNotificationModeRequest:
      type: object
      required: [mode]
      properties:
        room_id:
          type: string
          nullable: true
          description: Omitted/null sets the user default row; a room id sets that room's override.
        mode:
          type: string
          enum: [all, mentions, directs, never]
...
    MemberArchiveRequest:
      type: object
      required: [archived]
      properties:
        archived:
          type: boolean
...